+++ Fixtures/TargetWithCustomBuildRules/TargetWithCustomBuildRules/TargetWithCustomBuildRules.h
@@ -1,0 +1,12 @@
+//
+//  TargetWithCustomBuildRules.h
+//  TargetWithCustomBuildRules
+//
+//  Created by Pedro Piñera Buendía on 06.07.23.
+//
+
+#import <Foundation/Foundation.h>
+
+@interface TargetWithCustomBuildRules : NSObject
+
+@end