+++ Tests/XcodeProjTests/Tests/Fixtures.swift
@@ -15,3 +15,9 @@ func fileSharedAcrossTargetsDictionary() -> (Path, [String: Any]) {
     let fileSharedAcrossTargetsProject = fixturesPath() + "FileSharedAcrossTargets/FileSharedAcrossTargets.xcodeproj/project.pbxproj"
     return (fileSharedAcrossTargetsProject, loadPlist(path: fileSharedAcrossTargetsProject.string)!)
 }
+
+
+func targetWithCustomBuildRulesDictionary() -> (Path, [String: Any]) {
+    let targetWithCustomBuildRulesProject = fixturesPath() + "TargetWithCustomBuildRules/TargetWithCustomBuildRules.xcodeproj/project.pbxproj"
+    return (targetWithCustomBuildRulesProject, loadPlist(path: targetWithCustomBuildRulesProject.string)!)
+}